Sum of two numbers

+++ Sum.java
@@ -3,11 +3,11 @@ class Sum
 {
   public static void main(String args[])
   {
-   int a,b;
+   float a,b;
    Scanner sc = new Scanner(System.in);
    System.out.println("enter a and  b values");
-   a=sc.nextInt();
-   b=sc.nextInt();
+   a=sc.nextFloat();
+   b=sc.nextFloat();
    System.out.println(a+b);
   }
 }
